refactor(relation): extract state, DSL, options into modules
Split monolithic Relation into State, DSL, Options plus
Compiler/Materializers; slim facade, preserve API; wire tests helpers
and fix joins fixture; tests green

Closes: AE-161
