+++ lib/search_engine/relation.rb
@@ -3,82 +3,32 @@
 module SearchEngine
   # Immutable, chainable query relation bound to a model class.
   #
-  # Carries normalized query state and provides copy-on-write chainers.
-  # All chainers return new frozen instances; no in-place mutation occurs.
-  #
-  # @example Basic chaining
-  #   class Product < SearchEngine::Base; end
-  #   r1 = Product.all
-  #   r2 = r1.where(category: 'milk').order(:name).limit(10)
-  #   r1.object_id != r2.object_id #=> true
-  #   r1.empty?                    #=> true
+  # Facade wiring that composes State, Options, DSL, Compiler, and Materializers.
   class Relation
-    # Internal normalized state keys
-    DEFAULT_STATE = {
-      filters: [].freeze,
-      ast:     [].freeze, # Predicate AST nodes (authoritative)
-      orders:  [].freeze,
-      select:  [].freeze,
-      # Nested include_fields selection state
-      select_nested: {}.freeze,           # { assoc(Symbol) => [field(String), ...] }
-      select_nested_order: [].freeze,     # [assoc(Symbol), ...] first-mention order
-      # Exclude fields state (root + nested)
-      exclude: [].freeze,
-      exclude_nested: {}.freeze,          # { assoc(Symbol) => [field(String), ...] }
-      exclude_nested_order: [].freeze,    # [assoc(Symbol), ...] first-mention order
-      joins:   [].freeze,
-      limit:   nil,
-      offset:  nil,
-      page:    nil,
-      per_page: nil,
-      grouping: nil,
-      options: {}.freeze,
-      # Presets: immutable relation-level preset state
-      preset_name: nil,
-      preset_mode: nil,
-      facet_fields: [].freeze,
-      facet_max_values: [].freeze,
-      facet_queries: [].freeze,
-      highlight: {}.freeze,
-      highlight_fields: [].freeze,
-      highlight_full_fields: [].freeze,
-      highlight_start_tag: nil,
-      highlight_end_tag: nil,
-      highlight_affix_num_tokens: nil,
-      highlight_snippet_threshold: nil,
-      # Runtime flags: per-relation overrides for synonyms/stopwords
-      use_synonyms: nil,
-      use_stopwords: nil,
-      # Ranking/typo tuning (nil when unset)
-      ranking: nil,
-      hit_limits: {}.freeze
-    }.freeze
-
     # Keys considered essential for :only preset mode.
-    # Intentionally minimal: query basics and pagination; other keys are dropped.
     ESSENTIAL_PARAM_KEYS = %i[q page per_page].freeze
-
-    # Keys managed by preset in :lock mode. When present on the compiled params,
-    # they will be dropped and recorded as conflicts.
-    PRESET_MANAGED_PARAM_KEYS = %i[
-      filter_by
-      sort_by
-      include_fields
-      exclude_fields
-      facet_by
-      max_facet_values
-      group_by
-      group_limit
-      group_missing_values
-    ].freeze
 
     # @return [Class] bound model class (typically a SearchEngine::Base subclass)
     attr_reader :klass
 
+    # Modules are required explicitly to keep require graph stable
+    require 'search_engine/relation/state'
+    require 'search_engine/relation/options'
+    require 'search_engine/relation/dsl'
+    require 'search_engine/relation/compiler'
+    require 'search_engine/relation/materializers'
+
+    include State
+    include Options
+    include DSL
+    include Compiler
+    include Materializers
+
+    # Convenience alias to compiled body params.
+    alias to_h to_typesense_params
+
     # Read-only access to accumulated predicate AST nodes.
-    #
     # @return [Array<SearchEngine::AST::Node>] a frozen Array of AST nodes
-    # @note The returned array is frozen; modifying it will raise.
     def ast
       nodes = Array(@state[:ast])
       nodes.frozen? ? nodes : nodes.dup.freeze
@@ -86,25 +36,20 @@ module SearchEngine
 
     # Return the effective preset mode when a preset is applied.
     # Falls back to :merge when not explicitly set.
-    #
-    # @return [Symbol] one of :merge, :only, :lock
-    # @see docs/presets.md#relation-dsl
+    # @return [Symbol]
     def preset_mode
       (@state[:preset_mode] || :merge).to_sym
     end
 
     # Return the effective preset token (namespaced if configured) or nil.
-    #
     # @return [String, nil]
-    # @see docs/presets.md#config-default-preset
     def preset_name
       @state[:preset_name]
     end
 
     # Create a new Relation.
-    #
-    # @param klass [Class] model class the relation is bound to
-    # @param state [Hash] optional pre-populated normalized state
+    # @param klass [Class]
+    # @param state [Hash]
     def initialize(klass, state = {})
       @klass = klass
       normalized = normalize_initial_state(state)
@@ -120,556 +65,6 @@ module SearchEngine
     # @return [SearchEngine::Relation]
     def all
       self
-    end
-
-    # Add filters to the relation.
-    #
-    # Accepted forms:
-    # - Hash: where(id: 1, brand_id: [1,2,3])
-    # - Joined Hash: where(authors: { last_name: "Rowling" })
-    # - Raw string fragment: where("brand_id:=[1,2,3]")
-    # - Template with placeholders: where("price > ?", 100)
-    #
-    # Multiple calls compose with AND semantics (filters accumulate).
-    # The relation is immutable; a new instance is returned.
-    #
-    # @param args [Array<Object>] filter arguments
-    # @return [SearchEngine::Relation]
-    def where(*args)
-      # Build AST nodes for all supported inputs via the DSL parser
-      ast_nodes = SearchEngine::DSL::Parser.parse_list(args, klass: @klass, joins: joins_list)
-      # Back-compat: preserve legacy string fragments as well (escape hatch)
-      fragments = normalize_where(args)
-      spawn do |s|
-        s[:ast] = Array(s[:ast]) + Array(ast_nodes)
-        s[:filters] = Array(s[:filters]) + fragments
-      end
-    end
-
-    # Append ordering expressions. Accepts Hash or String forms.
-    #
-    # Accepted input:
-    # - Hash: { field => :asc|:desc, ... } or { assoc => { field => :asc|:desc } }
-    # - String: "field:dir" or comma-separated "field:dir,other:asc"
-    #
-    # Normalization:
-    # - Stored as array of strings like ["field:asc", "$assoc.field:desc"]
-    # - Direction lowercased; field trimmed; validation enforced
-    # - Dedupe by field with last-wins semantics while preserving last position
-    #
-    # @param value [Hash, String]
-    # @return [SearchEngine::Relation]
-    # @raise [ArgumentError] when direction or field is invalid
-    def order(value)
-      additions = normalize_order(value)
-      spawn do |s|
-        existing = Array(s[:orders])
-        s[:orders] = dedupe_orders_last_wins(existing + additions)
-      end
-    end
-
-    # Select a subset of fields for Typesense `include_fields`.
-    #
-    # Accepts a mix of base fields and nested association fields using a Ruby-ish shape.
-    #
-    # - Base fields: symbols/strings, e.g. `:id, "title"`
-    # - Nested: a Hash mapping association name => field list, e.g. `{ authors: [:first_name, :last_name] }`
-    # - Arrays are flattened; blanks are dropped; duplicates are removed preserving first occurrence order
-    # - Associations are validated against the model's `joins_config`
-    #
-    # Multiple calls merge:
-    # - Base fields: first-mention order wins; later calls append new unique fields only
-    # - Nested per-association: first-mention order wins; later calls append new unique fields only
-    # - Association emission order is by first mention across calls
-    #
-    # The relation is immutable; a new instance is returned.
-    #
-    # @param fields [Array<Symbol,String,Hash,Array>]
-    # @return [SearchEngine::Relation]
-    # @raise [ArgumentError] when inputs are invalid or field names are blank
-    # @raise [SearchEngine::Errors::UnknownJoin] when a nested association is not declared
-    def select(*fields)
-      normalized = normalize_select_input(fields)
-      spawn do |s|
-        # Merge base fields with first-wins semantics
-        existing_base = Array(s[:select])
-        merged_base = (existing_base + normalized[:base]).each_with_object([]) do |f, acc|
-          acc << f unless acc.include?(f)
-        end
-        s[:select] = merged_base
-
-        # Merge nested selections preserving association first-mention order
-        existing_nested = s[:select_nested] || {}
-        existing_order = Array(s[:select_nested_order])
-
-        normalized[:nested_order].each do |assoc|
-          new_fields = Array(normalized[:nested][assoc])
-          next if new_fields.empty?
-
-          old_fields = Array(existing_nested[assoc])
-          merged_fields = (old_fields + new_fields).each_with_object([]) do |name, acc|
-            acc << name unless acc.include?(name)
-          end
-          existing_nested = existing_nested.merge(assoc => merged_fields)
-          existing_order << assoc unless existing_order.include?(assoc)
-        end
-
-        s[:select_nested] = existing_nested
-        s[:select_nested_order] = existing_order
-      end
-    end
-
-    # Convenience alias for `select` supporting nested include_fields input.
-    #
-    # @see #select
-    # @param fields [Array]
-    # @return [SearchEngine::Relation]
-    def include_fields(*fields)
-      select(*fields)
-    end
-
-    # Apply a server-side preset with a specified merge strategy.
-    #
-    # Computes the effective preset name using global presets configuration.
-    # When presets are enabled and a namespace is present, prefixes the name
-    # with "#{SearchEngine.config.presets.namespace}_"; otherwise uses the
-    # provided token as a String.
-    #
-    # Modes:
-    # - :merge (default): emit preset and all chain params; chain wins on overlaps
-    # - :only: emit preset and only essential params (q, page, per_page)
-    # - :lock: emit preset and all chain params, but drop keys present in
-    #          SearchEngine.config.presets.locked_domains
-    #
-    # @param name [#to_s, #to_sym] preset token (without namespace)
-    # @param mode [Symbol] one of :merge, :only, :lock
-    # @return [SearchEngine::Relation]
-    # @raise [ArgumentError] on invalid name or mode
-    # @example
-    #   # Merge (default)
-    #   SearchEngine::Product.preset(:popular_products)
-    #     .where(active: true)
-    #     .order(updated_at: :desc)
-    #
-    #   # Only preset
-    #   SearchEngine::Product.preset(:aggressive_sale, mode: :only).page(1).per(24)
-    #
-    #   # Locked preset (chain cannot override preset filters/sorts)
-    #   SearchEngine::Product.preset(:brand_curated, mode: :lock).order(price: :asc) # order will be dropped
-    # @see docs/presets.md#relation-dsl
-    # @see docs/presets.md#strategies-merge-only-lock
-    def preset(name, mode: :merge)
-      raise ArgumentError, 'preset requires a name' if name.nil?
-
-      token = name.to_s.strip
-      raise ArgumentError, 'preset name must be non-empty' if token.empty?
-
-      sym_mode = mode.to_sym
-      unless %i[merge only lock].include?(sym_mode)
-        raise ArgumentError, "preset mode must be one of :merge, :only, :lock (got #{mode.inspect})"
-      end
-
-      cfg = SearchEngine.config.presets
-      effective = if cfg.enabled && cfg.namespace
-                    +"#{cfg.namespace}_#{token}"
-                  else
-                    token.dup
-                  end
-
-      spawn do |s|
-        s[:preset_name] = effective
-        s[:preset_mode] = sym_mode
-      end
-    end
-
-    # Fine-grained ranking & typo tuning. Merges immutably; last-writer wins for scalars,
-    # and weight hashes are merged with later keys overriding earlier ones.
-    #
-    # @param num_typos [Integer, String, nil] one of 0, 1, 2
-    # @param drop_tokens_threshold [Float, String, nil] 0.0..1.0 inclusive
-    # @param prioritize_exact_match [Boolean, String, Integer, nil] strict boolean
-    # @param query_by_weights [Hash{Symbol,String=>Integer,#to_i}, nil] weights per field (>= 0)
-    # @return [SearchEngine::Relation]
-    # @raise [SearchEngine::Errors::InvalidOption] when inputs are invalid
-    # @see docs/ranking.md
-    def ranking(num_typos: nil, drop_tokens_threshold: nil, prioritize_exact_match: nil, query_by_weights: nil)
-      normalized = normalize_ranking_input(
-        num_typos: num_typos,
-        drop_tokens_threshold: drop_tokens_threshold,
-        prioritize_exact_match: prioritize_exact_match,
-        query_by_weights: query_by_weights
-      )
-
-      spawn do |s|
-        current = s[:ranking] || {}
-        merged = current.dup
-        %i[num_typos drop_tokens_threshold prioritize_exact_match].each do |k|
-          merged[k] = normalized[k] unless normalized[k].nil?
-        end
-        if normalized.key?(:query_by_weights)
-          existing = current[:query_by_weights] || {}
-          merged[:query_by_weights] = existing.merge(normalized[:query_by_weights])
-        end
-        s[:ranking] = merged
-      end
-    end
-
-    # Control Typesense infix/prefix matching per relation via a simple enum.
-    # Accepts :disabled, :fallback, or :always. Maps to authoritative `infix` tokens
-    # ("off", "fallback", "always").
-    #
-    # @param mode [Symbol, String]
-    # @return [SearchEngine::Relation]
-    # @raise [SearchEngine::Errors::InvalidOption] when mode is unknown
-    # @since 0.1.0
-    # @example
-    #   SearchEngine::Product.prefix(:fallback)
-    def prefix(mode)
-      sym = mode.to_s.strip.downcase.to_sym
-      valid = {
-        disabled: 'off',
-        fallback: 'fallback',
-        always: 'always'
-      }
-      unless valid.key?(sym)
-        raise SearchEngine::Errors::InvalidOption.new(
-          "InvalidOption: unknown prefix mode #{mode.inspect}",
-          hint: 'Use :disabled, :fallback, or :always',
-          doc: 'docs/ranking.md#prefix',
-          details: { provided: mode, allowed: valid.keys }
-        )
-      end
-
-      token = valid[sym]
-      spawn do |s|
-        opts = (s[:options] || {}).dup
-        opts[:infix] = token
-        s[:options] = opts
-      end
-    end
-
-    # Exclude a subset of fields from the final selection.
-    #
-    # Accepts the same shapes as {#select}: base symbols/strings and nested hashes
-    # keyed by association => [fields]. Entries are normalized, deduplicated, and
-    # merged immutably into exclude stores.
-    #
-    # @param fields [Array<Symbol,String,Hash,Array>]
-    # @return [SearchEngine::Relation]
-    # @raise [ArgumentError] when inputs are invalid
-    # @see #select
-    def exclude(*fields)
-      normalized = normalize_select_input(fields, context: 'excluding fields')
-      spawn do |s|
-        # Base excludes: first-mention order preserved
-        existing_base = Array(s[:exclude])
-        merged_base = (existing_base + normalized[:base]).each_with_object([]) do |f, acc|
-          acc << f unless acc.include?(f)
-        end
-        s[:exclude] = merged_base
-
-        # Nested excludes: preserve assoc order by first mention
-        existing_nested = s[:exclude_nested] || {}
-        existing_order = Array(s[:exclude_nested_order])
-
-        normalized[:nested_order].each do |assoc|
-          new_fields = Array(normalized[:nested][assoc])
-          next if new_fields.empty?
-
-          old_fields = Array(existing_nested[assoc])
-          merged_fields = (old_fields + new_fields).each_with_object([]) do |name, acc|
-            acc << name unless acc.include?(name)
-          end
-          existing_nested = existing_nested.merge(assoc => merged_fields)
-          existing_order << assoc unless existing_order.include?(assoc)
-        end
-
-        s[:exclude_nested] = existing_nested
-        s[:exclude_nested_order] = existing_order
-      end
-    end
-
-    # Pin hits to the top of results by ID. Appends new IDs while de-duplicating
-    # and preserving first occurrence order.
-    #
-    # @param ids [Array<#to_s>] one or more IDs; arrays are flattened one level
-    # @return [SearchEngine::Relation]
-    # @see docs/curation.md
-    def pin(*ids)
-      additions = normalize_curation_ids(ids)
-      return self if additions.empty?
-
-      spawn do |s|
-        cur = s[:curation] || { pinned: [], hidden: [], override_tags: [], filter_curated_hits: nil }
-        cur[:pinned] = (Array(cur[:pinned]) + additions).each_with_object([]) do |t, acc|
-          acc << t unless acc.include?(t)
-        end
-        s[:curation] = cur
-      end
-    end
-
-    # Hide hits by ID. Appends new IDs while de-duplicating with set semantics
-    # and preserving first-seen order for display.
-    #
-    # @param ids [Array<#to_s>] one or more IDs; arrays are flattened one level
-    # @return [SearchEngine::Relation]
-    # @see docs/curation.md
-    def hide(*ids)
-      additions = normalize_curation_ids(ids)
-      return self if additions.empty?
-
-      spawn do |s|
-        cur = s[:curation] || { pinned: [], hidden: [], override_tags: [], filter_curated_hits: nil }
-        cur[:hidden] = (Array(cur[:hidden]) + additions).each_with_object([]) do |t, acc|
-          acc << t unless acc.include?(t)
-        end
-        s[:curation] = cur
-      end
-    end
-
-    # Set multiple curation knobs in one call. For provided keys, replaces the
-    # corresponding list/value; omitted keys retain previous values.
-    #
-    # @param pin [Array<#to_s>, nil]
-    # @param hide [Array<#to_s>, nil]
-    # @param override_tags [Array<#to_s>, nil]
-    # @param filter_curated_hits [true,false,nil,Symbol]
-    # @return [SearchEngine::Relation]
-    # @see docs/curation.md
-    def curate(pin: nil, hide: nil, override_tags: nil, filter_curated_hits: :__unset__)
-      spawn do |s|
-        cur = s[:curation] || { pinned: [], hidden: [], override_tags: [], filter_curated_hits: nil }
-
-        unless pin.nil?
-          list = normalize_curation_ids(pin)
-          cur[:pinned] = list.each_with_object([]) { |t, acc| acc << t unless acc.include?(t) }
-        end
-        unless hide.nil?
-          list = normalize_curation_ids(hide)
-          cur[:hidden] = list.each_with_object([]) { |t, acc| acc << t unless acc.include?(t) }
-        end
-        cur[:override_tags] = normalize_curation_tags(override_tags) unless override_tags.nil?
-        if filter_curated_hits != :__unset__
-          cur[:filter_curated_hits] =
-            filter_curated_hits.nil? ? nil : coerce_boolean_strict(filter_curated_hits, :filter_curated_hits)
-        end
-
-        s[:curation] = cur
-      end
-    end
-
-    # Clear all curation state from the relation.
-    # @return [SearchEngine::Relation]
-    # @see docs/curation.md
-    def clear_curation
-      spawn do |s|
-        s[:curation] = nil
-      end
-    end
-
-    # Group results by a single field with optional limit and missing values policy.
-    #
-    # Stores normalized immutable grouping state under `@state[:grouping]`.
-    # Subsequent calls replace the existing grouping state (last call wins).
-    #
-    # @param field [Symbol, String] field name to group by (base field only)
-    # @param limit [Integer, nil] optional positive limit for number of hits per group
-    # @param missing_values [Boolean] whether to include missing values as their own group
-    # @return [SearchEngine::Relation] a new relation with grouping applied
-    # @raise [SearchEngine::Errors::InvalidGroup] when inputs are invalid or field is unknown
-    # @raise [SearchEngine::Errors::UnsupportedGroupField] when a joined/path field is provided (unsupported)
-    # @example
-    #   SearchEngine::Product
-    #     .group_by(:brand_id, limit: 1, missing_values: true)
-    #     .where(active: true)
-    #     .order(updated_at: :desc)
-    def group_by(field, limit: nil, missing_values: false)
-      normalized = normalize_grouping(field: field, limit: limit, missing_values: missing_values)
-
-      rel = spawn do |s|
-        s[:grouping] = normalized
-      end
-
-      if defined?(SearchEngine::Instrumentation)
-        begin
-          payload = {
-            collection: klass_name_for_inspect,
-            field: normalized[:field].to_s,
-            limit: normalized[:limit],
-            missing_values: normalized[:missing_values]
-          }
-          SearchEngine::Instrumentation.instrument('search_engine.relation.group_by_updated', payload) {}
-        rescue StandardError
-          # swallow observability errors
-        end
-      end
-
-      rel
-    end
-
-    # Replace the selected fields list (Typesense `include_fields`).
-    #
-    # Immutably replaces the existing selection with the normalized list of
-    # field names. Values are flattened, stripped, coerced to String, blanks
-    # dropped, and duplicates removed (preserving first occurrence).
-    #
-    # @param fields [Array<#to_sym,#to_s>] one or more field identifiers
-    # @return [SearchEngine::Relation] a new relation with replaced selection
-    # @raise [ArgumentError] when the resulting list is empty, a field is blank,
-    #   or a field is unknown for the model when attributes are declared.
-    # @example
-    #   rel.reselect(:id, :name)
-    def reselect(*fields)
-      normalized = normalize_select_input(fields)
-
-      base_empty = Array(normalized[:base]).empty?
-      nested_empty = normalized[:nested_order].all? { |a| Array(normalized[:nested][a]).empty? }
-      raise ArgumentError, 'reselect: provide at least one non-blank field' if base_empty && nested_empty
-
-      spawn do |s|
-        s[:select] = normalized[:base]
-        s[:select_nested] = normalized[:nested]
-        s[:select_nested_order] = normalized[:nested_order]
-        # Clear excludes when reselecting (explicit reset)
-        s[:exclude] = []
-        s[:exclude_nested] = {}
-        s[:exclude_nested_order] = []
-      end
-    end
-
-    # Replace all predicates with a new where input.
-    #
-    # Clears prior predicate state (AST and legacy filters) and parses
-    # the provided input into a fresh AST using the DSL parser.
-    # Accepts the same input forms as {#where}.
-    #
-    # @param input [Hash, String, Array, Symbol] predicate input
-    # @param args [Array<Object>] arguments for template strings
-    # @return [SearchEngine::Relation] a new relation with replaced predicates
-    # @raise [ArgumentError] when input is missing/blank or produces no predicates
-    # @example
-    #   rel.rewhere(active: true)
-    def rewhere(input, *args)
-      if input.nil? || (input.respond_to?(:empty?) && input.empty?) || (input.is_a?(String) && input.strip.empty?)
-        raise ArgumentError, 'rewhere: provide a new predicate input'
-      end
-
-      nodes = SearchEngine::DSL::Parser.parse(input, klass: @klass, args: args, joins: joins_list)
-      list = Array(nodes).flatten.compact
-      raise ArgumentError, 'rewhere: produced no predicates' if list.empty?
-
-      spawn do |s|
-        s[:ast] = list
-        s[:filters] = []
-      end
-    end
-
-    # Remove specific pieces of relation state (AR-style unscope).
-    #
-    # Supported parts and their effects:
-    # - :where  => clears AST and legacy filters
-    # - :order  => clears orders
-    # - :select => clears field selection
-    # - :limit  => sets limit to nil
-    # - :offset => sets offset to nil
-    # - :page   => sets page to nil
-    # - :per    => sets per_page to nil
-    #
-    # @param parts [Array<Symbol,String>] one or more parts to remove
-    # @return [SearchEngine::Relation] a new relation with state removed
-    # @raise [ArgumentError] when an unsupported part is provided
-    # @example
-    #   rel.unscope(:order)
-    def unscope(*parts)
-      symbols = Array(parts).flatten.compact.map(&:to_sym)
-      supported = %i[where order select limit offset page per]
-      unknown = symbols - supported
-      unless unknown.empty?
-        raise ArgumentError,
-              "unscope: unknown part #{unknown.first.inspect} (supported: #{supported.map(&:inspect).join(', ')})"
-      end
-
-      spawn do |s|
-        symbols.each do |part|
-          case part
-          when :where
-            s[:ast] = []
-            s[:filters] = []
-          when :order
-            s[:orders] = []
-          when :select
-            s[:select] = []
-            s[:select_nested] = {}
-            s[:select_nested_order] = []
-            s[:exclude] = []
-            s[:exclude_nested] = {}
-            s[:exclude_nested_order] = []
-          when :limit
-            s[:limit] = nil
-          when :offset
-            s[:offset] = nil
-          when :page
-            s[:page] = nil
-          when :per
-            s[:per_page] = nil
-          end
-        end
-      end
-    end
-
-    # Set the maximum number of results.
-    # @param n [Integer, #to_i, nil]
-    # @return [SearchEngine::Relation]
-    # @raise [ArgumentError] when n < 1 or not coercible to Integer
-    def limit(n)
-      value = coerce_integer_min(n, :limit, 1)
-      spawn { |s| s[:limit] = value }
-    end
-
-    # Set the offset of results.
-    # @param n [Integer, #to_i, nil]
-    # @return [SearchEngine::Relation]
-    # @raise [ArgumentError] when n < 0 or not coercible to Integer
-    def offset(n)
-      value = coerce_integer_min(n, :offset, 0)
-      spawn { |s| s[:offset] = value }
-    end
-
-    # Set page number.
-    # @param n [Integer, #to_i, nil]
-    # @return [SearchEngine::Relation]
-    # @raise [ArgumentError] when n < 1 or not coercible to Integer
-    def page(n)
-      value = coerce_integer_min(n, :page, 1)
-      spawn { |s| s[:page] = value }
-    end
-
-    # Set per-page size.
-    # @param n [Integer, #to_i, nil]
-    # @return [SearchEngine::Relation]
-    # @raise [ArgumentError] when n < 1 or not coercible to Integer
-    def per_page(n)
-      value = coerce_integer_min(n, :per, 1)
-      spawn { |s| s[:per_page] = value }
-    end
-
-    # Convenience alias for per-page size.
-    # @param n [Integer, #to_i, nil]
-    # @return [SearchEngine::Relation]
-    def per(n)
-      per_page(n)
-    end
-
-    # Shallow-merge options into the relation.
-    # For nested hashes, merging is shallow by default.
-    # @param opts [Hash]
-    # @return [SearchEngine::Relation]
-    def options(opts = {})
-      raise ArgumentError, 'options must be a Hash' unless opts.is_a?(Hash)
-
-      spawn do |s|
-        s[:options] = (s[:options] || {}).merge(opts)
-      end
     end
 
     # True when the relation has no accumulated state beyond defaults.
@@ -719,302 +114,9 @@ module SearchEngine
       "#<#{self.class.name} #{parts.join(' ')} >"
     end
 
-    # Compile immutable relation state and options into Typesense body params.
-    # The compiler is pure/deterministic and omits URL-level cache knobs.
-    #
-    # Key insertion order: q, query_by, filter_by, sort_by, include_fields, page, per_page, infix.
-    # Empty/nil values are omitted.
-    #
-    # May emit grouping keys when present:
-    # - { group_by, group_limit, group_missing_values }
-    #
-    # Presets:
-    # - Always injects `preset` when present on the relation
-    # - mode=:merge — no pruning
-    # - mode=:only  — keep only essentials (q, page, per_page)
-    # - mode=:lock  — prune keys in SearchEngine.config.presets.locked_domains
-    # @return [Hash] typesense body params suitable for Client#search
-    # @example
-    #   rel.to_typesense_params
-    #   # => { q: "*", query_by: "name,description", filter_by: "brand_id:=[1,2] && active:=true", page: 2, per_page: 20 }
-    # @see docs/presets.md#compiler-mapping-pruning
-    def to_typesense_params # rubocop:disable Metrics/MethodLength, Metrics/AbcSize, Metrics/PerceivedComplexity
-      cfg = SearchEngine.config
-      opts = @state[:options] || {}
-
-      params = {}
-      runtime_flags = {}
-
-      # Query basics
-      q_val = option_value(opts, :q) || '*'
-      query_by_val = option_value(opts, :query_by) || cfg.default_query_by
-      params[:q] = q_val
-      params[:query_by] = query_by_val if query_by_val
-
-      # Filters and sorting
-      ast_nodes = Array(@state[:ast]).flatten.compact
-      SearchEngine::Instrumentation.monotonic_ms if defined?(SearchEngine::Instrumentation)
-      filter_str = compiled_filter_by(ast_nodes)
-      params[:filter_by] = filter_str if filter_str
-
-      orders = Array(@state[:orders])
-      sort_str = compiled_sort_by(orders)
-      params[:sort_by] = sort_str if sort_str
-
-      # Field selection (nested first, then base), honoring excludes
-      include_str = compile_include_fields_string
-      params[:include_fields] = include_str unless include_str.to_s.strip.empty?
-
-      # Exclude fields when provided and not fully derivable via include
-      exclude_str = compile_exclude_fields_string
-      params[:exclude_fields] = exclude_str unless exclude_str.to_s.strip.empty?
-
-      # Highlighting
-      if (h = @state[:highlight])
-        hf = Array(h[:fields]).map(&:to_s).reject(&:empty?)
-        params[:highlight_fields] = hf.join(',') unless hf.empty?
-
-        hff = Array(h[:full_fields]).map(&:to_s).reject(&:empty?)
-        params[:highlight_full_fields] = hff.join(',') unless hff.empty?
-
-        params[:highlight_start_tag] = h[:start_tag] if h[:start_tag]
-        params[:highlight_end_tag] = h[:end_tag] if h[:end_tag]
-
-        params[:highlight_affix_num_tokens] = h[:affix_tokens] unless h[:affix_tokens].nil?
-        params[:snippet_threshold] = h[:snippet_threshold] unless h[:snippet_threshold].nil?
-      end
-
-      # Faceting
-      facet_fields = Array(@state[:facet_fields]).map(&:to_s).reject(&:empty?)
-      params[:facet_by] = facet_fields.join(',') unless facet_fields.empty?
-
-      caps = Array(@state[:facet_max_values]).compact
-      if caps.any?
-        valid_caps = []
-        caps.each do |v|
-          valid_caps << Integer(v)
-        rescue ArgumentError, TypeError
-          # skip invalid cap
-        end
-        max_cap = valid_caps.max
-        params[:max_facet_values] = max_cap if max_cap&.positive?
-      end
-
-      queries = Array(@state[:facet_queries])
-      if queries.any?
-        tokens = queries.map { |q| "#{q[:field]}:#{q[:expr]}" }
-        params[:facet_query] = tokens.join(',') unless tokens.empty?
-      end
-
-      # Curation (body params only)
-      if (cur = @state[:curation])
-        pinned = Array(cur[:pinned]).map(&:to_s).reject(&:empty?)
-        hidden = Array(cur[:hidden]).map(&:to_s).reject(&:empty?)
-        tags   = Array(cur[:override_tags]).map(&:to_s).reject(&:empty?)
-        fch    = cur[:filter_curated_hits]
-
-        params[:pinned_hits] = pinned.join(',') if pinned.any?
-        params[:hidden_hits] = hidden.join(',') if hidden.any?
-        params[:override_tags] = tags.join(',') if tags.any?
-        params[:filter_curated_hits] = fch unless fch.nil?
-
-        # Emit curation compile event (counts/flags only)
-        if defined?(SearchEngine::Instrumentation)
-          begin
-            c_payload = {
-              pinned_count: pinned.size.positive? ? pinned.size : nil,
-              hidden_count: hidden.size.positive? ? hidden.size : nil,
-              has_override_tags: tags.any? || nil,
-              filter_curated_hits: (cur.key?(:filter_curated_hits) ? fch : nil)
-            }.compact
-            SearchEngine::Instrumentation.instrument('search_engine.curation.compile', c_payload) {}
-
-            overlap = (pinned & hidden)
-            if overlap.any?
-              SearchEngine::Instrumentation.instrument(
-                'search_engine.curation.conflict',
-                { type: :overlap, count: overlap.size }
-              ) {}
-            end
-          rescue StandardError
-            # swallow observability errors
-          end
-        end
-      end
-
-      # Pagination
-      pagination = compute_pagination
-      # Early limit (compiler mapping): Typesense has no canonical total-hits cap.
-      # Apply conservative fallback by adjusting per_page downward when it exceeds the limit.
-      hits_info = {}
-      if (hl = @state[:hit_limits]) && hl[:early_limit]
-        if pagination.key?(:per_page) && pagination[:per_page].to_i > hl[:early_limit].to_i
-          pagination = pagination.merge(per_page: hl[:early_limit].to_i)
-          hits_info[:per_adjusted] = true
-        else
-          hits_info[:per_adjusted] = false
-        end
-        hits_info[:early_limit] = hl[:early_limit].to_i
-      end
-      params[:page] = pagination[:page] if pagination.key?(:page)
-      params[:per_page] = pagination[:per_page] if pagination.key?(:per_page)
-
-      # Grouping
-      grouping = @state[:grouping]
-      if grouping
-        field = grouping[:field]
-        limit = grouping[:limit]
-        missing_values = grouping[:missing_values]
-
-        if field
-          params[:group_by] = field.to_s
-          params[:group_limit] = limit if limit
-          params[:group_missing_values] = true if missing_values
-        end
-
-        if defined?(SearchEngine::Instrumentation)
-          begin
-            payload = {
-              collection: klass_name_for_inspect,
-              field: field&.to_s,
-              limit: limit,
-              missing_values: missing_values
-            }.compact
-            SearchEngine::Instrumentation.instrument('search_engine.grouping.compile', payload) {}
-          rescue StandardError
-            # swallow observability errors
-          end
-        end
-      end
-
-      # Keep infix last for stability; include when configured or overridden
-      infix_val = option_value(opts, :infix) || cfg.default_infix
-      params[:infix] = infix_val if infix_val
-
-      # Ranking & typo tuning — authoritative mapping
-      if (rk = @state[:ranking])
-        begin
-          plan = SearchEngine::RankingPlan.new(relation: self, query_by: params[:query_by], ranking: rk)
-          rparams = plan.params
-          params.merge!(rparams) unless rparams.empty?
-        rescue SearchEngine::Errors::Error
-          raise
-        rescue StandardError => error
-          raise SearchEngine::Errors::InvalidOption.new(
-            "InvalidOption: ranking options could not be compiled (#{error.class}: #{error.message})",
-            doc: 'docs/ranking.md#options'
-          )
-        end
-      end
-
-      # Internal join context (for downstream components; may be stripped before HTTP)
-      join_ctx = build_join_context(ast_nodes: ast_nodes, orders: orders)
-      params[:_join] = join_ctx unless join_ctx.nil? || join_ctx.empty?
-
-      # Preset emission and merge strategies
-      if (pn = @state[:preset_name])
-        pmode = (@state[:preset_mode] || :merge).to_sym
-        params[:preset] = pn
-
-        case pmode
-        when :only
-          allowed = ESSENTIAL_PARAM_KEYS
-          minimal = {}
-          (allowed + [:preset]).each do |k|
-            minimal[k] = params[k] if params.key?(k)
-          end
-          # Preserve internal join context if present for observability
-          minimal[:_join] = params[:_join] if params.key?(:_join)
-          params = minimal
-        when :lock
-          conflicts = []
-          locked = SearchEngine.config.presets.locked_domains_set
-          params.each_key do |k|
-            next unless locked.include?(k)
-
-            params.delete(k)
-            conflicts << k
-          end
-          params[:_preset_conflicts] = conflicts unless conflicts.empty?
-
-          instrument_preset_conflicts(pmode, pn, conflicts)
-        end
-      end
-
-      # Emit compile-time JOINs event summarizing usage (no raw strings)
-      if defined?(SearchEngine::Instrumentation)
-        begin
-          assocs = Array(join_ctx[:assocs]).map(&:to_s)
-          used = join_ctx[:referenced_in] || {}
-          used_in = {}
-          %i[include filter sort].each do |k|
-            arr = Array(used[k]).map(&:to_s)
-            used_in[k] = arr unless arr.empty?
-          end
-
-          payload = {
-            collection: klass_name_for_inspect,
-            join_count: assocs.size,
-            assocs: (assocs unless assocs.empty?),
-            used_in: (used_in unless used_in.empty?),
-            include_len: (include_str.to_s.length unless include_str.to_s.strip.empty?),
-            filter_len: (filter_str.to_s.length unless filter_str.to_s.strip.empty?),
-            sort_len:   (sort_str.to_s.length unless sort_str.to_s.strip.empty?),
-            duration_ms: (SearchEngine::Instrumentation.monotonic_ms - compile_started_ms if compile_started_ms),
-            has_joins: !assocs.empty?
-          }
-          SearchEngine::Instrumentation.instrument('search_engine.joins.compile', payload)
-        rescue StandardError
-          # swallow observability errors
-        end
-      end
-
-      # Synonyms/Stopwords toggles — map to authoritative server params.
-      # If server uses inverse semantics for stopwords (e.g., remove_stop_words),
-      # flip here while keeping the DSL stable.
-      unless @state[:use_synonyms].nil?
-        params[:enable_synonyms] = @state[:use_synonyms]
-        runtime_flags[:use_synonyms] = @state[:use_synonyms]
-      end
-      unless @state[:use_stopwords].nil?
-        # Typesense expects remove_stop_words=true to disable stopwords; invert our DSL if needed.
-        # Authoritative mapping assumption: param name is :remove_stop_words.
-        remove = !@state[:use_stopwords]
-        params[:remove_stop_words] = remove
-        runtime_flags[:use_stopwords] = @state[:use_stopwords]
-      end
-
-      # Attach internal-only runtime flags preview for observability; stripped in client
-      params[:_runtime_flags] = runtime_flags unless runtime_flags.empty?
-
-      # Attach internal-only hit limits preview for DX surfaces; stripped client-side before HTTP
-      if (hl = @state[:hit_limits])
-        hits_info[:max] = hl[:max].to_i if hl[:max]
-        params[:_hits] = hits_info unless hits_info.empty?
-      end
-
-      params
-    end
-
-    # Convenience alias to compiled body params.
-    alias to_h to_typesense_params
-
     # Explain the current relation without performing any network calls.
-    #
-    # Returns a concise, multi-line String summarizing chainers and compiled
-    # parameters. When `to: :stdout` is provided, also prints the summary.
-    # Includes overview (collection, page/per_page, preset, grouping), selection,
-    # joins count, curation flags, conflicts (e.g., preset :lock drops), and a
-    # predictive list of events that would fire. Correlation ID is previewed via
-    # instrumentation when available.
-    #
-    # @param to [Symbol, nil] when `:stdout`, prints to STDOUT in addition to returning
     # @return [String]
-    # @example
-    #   rel.explain
-    #   #=> "SearchEngine::Product Relation\n  where: active:=true AND brand_id IN [1,2]\n  order: updated_at:desc\n  select: id,name\n  page/per: 2/20"
-    # @see docs/presets.md#conflicts
-    def explain(to: nil) # rubocop:disable Metrics/AbcSize, Metrics/PerceivedComplexity
+    def explain(to: nil)
       params = to_typesense_params
 
       lines = []
@@ -1022,9 +124,9 @@ module SearchEngine
       lines << header
 
       append_preset_explain_line(lines, params)
+
       append_curation_explain_lines(lines)
 
-      # Runtime flags explain
       lines << "  use_synonyms: #{@state[:use_synonyms]}" if @state.key?(:use_synonyms) && !@state[:use_synonyms].nil?
       if @state.key?(:use_stopwords) && !@state[:use_stopwords].nil?
         lines << "  use_stopwords: #{@state[:use_stopwords]} (maps to remove_stop_words=#{!@state[:use_stopwords]})"
@@ -1045,21 +147,8 @@ module SearchEngine
       end
 
       append_selection_explain_lines(lines, params)
-      add_effective_selection_tokens!(lines)
 
-      # Highlight explain
-      if (h = @state[:highlight])
-        parts = []
-        f = Array(h[:fields]).map(&:to_s).reject(&:empty?)
-        parts << "fields=#{f.join(',')}" unless f.empty?
-        ff = Array(h[:full_fields]).map(&:to_s).reject(&:empty?)
-        parts << "full=#{ff.join(',')}" unless ff.empty?
-        parts << "start_tag=#{h[:start_tag]}" if h[:start_tag]
-        parts << "end_tag=#{h[:end_tag]}" if h[:end_tag]
-        parts << "affix_tokens=#{h[:affix_tokens]}" unless h[:affix_tokens].nil?
-        parts << "snippet_threshold=#{h[:snippet_threshold]}" unless h[:snippet_threshold].nil?
-        lines << "  highlight: #{parts.join(' ')}" unless parts.empty?
-      end
+      add_effective_selection_tokens!(lines)
 
       add_pagination_line!(lines, params)
 
@@ -1068,222 +157,15 @@ module SearchEngine
       out
     end
 
-    # Materializers
-    # --------------
-    #
-    # Each materializer triggers at most one HTTP request per Relation instance
-    # by memoizing the loaded Result. Subsequent calls reuse the memo.
-
-    # Return a shallow copy of hydrated hits.
-    # @return [Array<Object>]
-    def to_a
-      ensure_loaded!
-      @__result_memo.to_a
-    end
-
-    # Iterate over hydrated hits.
-    # @yieldparam obj [Object] hydrated object
-    # @return [Enumerator] when no block is given
-    def each(&block)
-      ensure_loaded!
-      return @__result_memo.each unless block_given?
-
-      @__result_memo.each(&block)
-    end
-
-    # Return the first element or the first N elements from the loaded page.
-    # When +n+ is provided, returns an Array; otherwise returns a single object or nil.
-    # @param n [Integer, nil]
-    # @return [Object, Array<Object>]
-    def first(n = nil)
-      ensure_loaded!
-      return @__result_memo.to_a.first if n.nil?
-
-      @__result_memo.to_a.first(n)
-    end
-
-    # Return the last element or the last N elements from the currently fetched page.
-    # Note: operates on the loaded page only; does not trigger additional requests.
-    # @param n [Integer, nil]
-    # @return [Object, Array<Object>]
-    def last(n = nil)
-      ensure_loaded!
-      return @__result_memo.to_a.last if n.nil?
-
-      @__result_memo.to_a.last(n)
-    end
-
-    # Take N elements from the head. When N==1, returns a single object.
-    # @param n [Integer]
-    # @return [Object, Array<Object>]
-    def take(n = 1)
-      ensure_loaded!
-      return @__result_memo.to_a.first if n == 1
-
-      @__result_memo.to_a.first(n)
-    end
-
-    # Convenience for plucking :id values.
-    # @return [Array<Object>]
-    def ids
-      pluck(:id)
-    end
-
-    # Pluck one or multiple fields. For a single field, returns a flat Array.
-    # For multiple fields, returns an Array of Arrays, preserving field order.
-    # Prefers calling readers on hydrated objects when available, otherwise falls
-    # back to reading directly from the raw document.
-    # @param fields [Array<#to_sym,#to_s>]
-    # @return [Array<Object>, Array<Array<Object>>]
-    def pluck(*fields)
-      raise ArgumentError, 'pluck requires at least one field' if fields.nil? || fields.empty?
-
-      # Validate against effective root selection BEFORE loading
-      names = coerce_pluck_field_names(fields)
-      validate_pluck_fields_allowed!(names)
-
-      ensure_loaded!
-      raw_hits = Array(@__result_memo.raw['hits'])
-      objects = @__result_memo.to_a
-
-      if names.length == 1
-        field = names.first
-        return objects.each_with_index.map do |obj, idx|
-          if obj.respond_to?(field)
-            obj.public_send(field)
-          else
-            doc = (raw_hits[idx] && raw_hits[idx]['document']) || {}
-            doc[field]
-          end
-        end
-      end
-
-      # Multiple fields -> array-of-arrays
-      objects.each_with_index.map do |obj, idx|
-        doc = (raw_hits[idx] && raw_hits[idx]['document']) || {}
-        names.map do |field|
-          if obj.respond_to?(field)
-            obj.public_send(field)
-          else
-            doc[field]
-          end
-        end
-      end
-    end
-
-    # Build a concise InvalidSelection message with prescriptive guidance.
-    # - Prefer `reselect` suggestion when includes are present and the field isn't excluded
-    # - Prefer `Remove exclude(:field)` when the field is explicitly excluded
-    def build_invalid_selection_message_for_pluck(missing:, requested:, include_base:, exclude_base:)
-      field = missing.map(&:to_s).min
-      if exclude_base.include?(field)
-        "InvalidSelection: field :#{field} not in effective selection. Remove exclude(:#{field})."
-      else
-        # Suggest reselect with current includes plus missing requested, de-duped and stable
-        suggestion_fields = include_base.dup
-        requested.each { |f| suggestion_fields << f unless suggestion_fields.include?(f) }
-        symbols = suggestion_fields.map { |t| ":#{t}" }.join(',')
-        "InvalidSelection: field :#{field} not in effective selection. Use `reselect(#{symbols})`."
-      end
-    end
-
-    # Normalize field inputs for pluck to deterministic string tokens.
-    def coerce_pluck_field_names(fields)
-      Array(fields).flatten.compact.map(&:to_s).map(&:strip).reject(&:empty?)
-    end
-
-    # Validate requested names against effective root selection and raise when any are missing.
-    def validate_pluck_fields_allowed!(names)
-      include_base = Array(@state[:select]).map(&:to_s)
-      exclude_base = Array(@state[:exclude]).map(&:to_s)
-
-      missing =
-        if include_base.empty?
-          # All fields allowed except explicit excludes
-          names & exclude_base
-        else
-          allowed = include_base - exclude_base
-          names - allowed
-        end
-
-      return if missing.empty?
-
-      msg = build_invalid_selection_message_for_pluck(
-        missing: missing,
-        requested: names,
-        include_base: include_base,
-        exclude_base: exclude_base
-      )
-      field = missing.map(&:to_s).min
-      hint = exclude_base.include?(field) ? "Remove exclude(:#{field})." : nil
-      raise SearchEngine::Errors::InvalidSelection.new(
-        msg,
-        hint: hint,
-        doc: 'docs/field_selection.md#guardrails',
-        details: { requested: names, include_base: include_base, exclude_base: exclude_base }
-      )
-    end
-
-    # Return total number of matching documents.
-    # Uses memoized Result when loaded; otherwise performs a minimal found-only request.
-    # @return [Integer]
-    def count
-      if curation_filter_curated_hits?
-        ensure_loaded!
-        return curated_indices_for_current_result.size
-      end
-
-      return @__result_memo.found.to_i if @__loaded && @__result_memo
-
-      fetch_found_only
-    end
-
-    # Whether any matching documents exist.
-    # Uses memoized Result when loaded; otherwise performs a minimal found-only request.
-    # @return [Boolean]
-    def exists?
-      return @__result_memo.found.to_i.positive? if @__loaded && @__result_memo
-
-      fetch_found_only.positive?
-    end
-
-    # Join association names to include in server-side join compilation.
-    #
-    # Validates each provided association against the model's `joins_config` and
-    # appends the normalized names to the relation state in order. The relation
-    # is immutable; a new instance is returned with copy-on-write semantics.
-    #
-    # @param assocs [Array<#to_sym,#to_s>] one or more association names
-    # @return [SearchEngine::Relation]
-    # @raise [SearchEngine::Errors::UnknownJoin] when an association is not declared
-    # @raise [ArgumentError] when inputs are not Symbols/Strings
-    # @see docs/joins.md
-    # @example
-    #   SearchEngine::Book
-    #     .joins(:authors)
-    #     .where(orders: { total_price: 12.34 })
-    def joins(*assocs)
-      names = normalize_joins(assocs)
-      return self if names.empty?
-
-      # Validate all first for better error UX
-      names.each { |name| SearchEngine::Joins::Guard.ensure_config_complete!(@klass, name) }
-
-      spawn do |s|
-        existing = Array(s[:joins])
-        s[:joins] = existing + names
-      end
-    end
-
     # Read-only list of join association names accumulated on this relation.
-    # @return [Array<Symbol>] a frozen array
+    # @return [Array<Symbol>]
     def joins_list
       list = Array(@state[:joins])
       list.frozen? ? list : list.dup.freeze
     end
 
     # Read-only grouping state for debugging/explain.
-    # @return [Hash, nil] frozen hash `{ field: Symbol, limit: Integer/nil, missing_values: Boolean }` or nil
+    # @return [Hash, nil]
     def grouping
       g = @state[:grouping]
       return nil if g.nil?
@@ -1306,9 +188,7 @@ module SearchEngine
     end
 
     # Programmatic accessor for preset conflicts in :lock mode.
-    # Builds a deterministic, frozen list of conflict entries. Keys only; no values.
-    # @return [Array<Hash{Symbol=>Symbol}>] e.g., [{ key: :sort_by, reason: :locked_by_preset }]
-    # @see docs/presets.md#conflicts
+    # @return [Array<Hash{Symbol=>Symbol}>]
     def preset_conflicts
       params = to_typesense_params
       keys = Array(params[:_preset_conflicts]).map { |k| k.respond_to?(:to_sym) ? k.to_sym : k }.grep(Symbol)
@@ -1318,7 +198,7 @@ module SearchEngine
     end
 
     # Read-only hit limits state for debugging/explain.
-    # @return [Hash, nil] frozen hash { early_limit: Integer/nil, max: Integer/nil } or nil
+    # @return [Hash, nil]
     def hit_limits
       hl = @state[:hit_limits]
       return nil if hl.nil?
@@ -1329,8 +209,6 @@ module SearchEngine
     protected
 
     # Spawn a new relation with a deep-duplicated mutable state.
-    # The given block may mutate the provided state Hash (shallow mutations only).
-    # Returns a new frozen Relation.
     # @yieldparam state [Hash]
     # @return [SearchEngine::Relation]
     def spawn
@@ -1343,475 +221,6 @@ module SearchEngine
 
     def klass_name_for_inspect
       @klass.respond_to?(:name) && @klass.name ? @klass.name : @klass.to_s
-    end
-
-    def format_value_for_inspect(value)
-      case value
-      when String
-        value.inspect
-      when Array
-        "[#{value.map { |v| format_value_for_inspect(v) }.join(', ')}]"
-      else
-        value.inspect
-      end
-    end
-
-    def friendly_where(filter_by)
-      s = filter_by.to_s
-      return s if s.empty?
-
-      # Replace tokens for readability
-      s = s.gsub(' && ', ' AND ')
-      s = s.gsub(' || ', ' OR ')
-      s = s.gsub(':=[', ' IN [')
-      s.gsub(':!=[', ' NOT IN [')
-    end
-
-    def normalize_initial_state(state)
-      return {} if state.nil? || state.empty?
-      raise ArgumentError, 'state must be a Hash' unless state.is_a?(Hash)
-
-      normalized = {}
-      state.each { |key, value| apply_initial_state_key!(normalized, key, value) }
-      normalized
-    end
-
-    def apply_initial_state_key!(normalized, key, value) # rubocop:disable Metrics/AbcSize, Metrics/PerceivedComplexity, Metrics/MethodLength
-      k = key.to_sym
-      case k
-      when :filters
-        normalized[:filters] = normalize_where(Array(value))
-      when :filters_ast
-        # Back-compat: accept legacy key and map to :ast
-        nodes = Array(value).flatten.compact
-        normalized[:ast] ||= []
-        normalized[:ast] += if nodes.all? { |n| n.is_a?(SearchEngine::AST::Node) }
-                              nodes
-                            else
-                              SearchEngine::DSL::Parser.parse_list(nodes, klass: @klass)
-                            end
-      when :ast
-        nodes = Array(value).flatten.compact
-        normalized[:ast] = if nodes.all? { |n| n.is_a?(SearchEngine::AST::Node) }
-                             nodes
-                           else
-                             SearchEngine::DSL::Parser.parse_list(nodes, klass: @klass)
-                           end
-      when :orders
-        normalized[:orders] = normalize_order(value)
-      when :select
-        normalized[:select] = normalize_select(Array(value))
-      when :select_nested
-        normalized[:select_nested] = (value || {})
-      when :select_nested_order
-        normalized[:select_nested_order] = Array(value).flatten.compact.map(&:to_sym)
-      when :exclude
-        normalized[:exclude] = normalize_select(Array(value))
-      when :exclude_nested
-        normalized[:exclude_nested] = (value || {})
-      when :exclude_nested_order
-        normalized[:exclude_nested_order] = Array(value).flatten.compact.map(&:to_sym)
-      when :joins
-        normalized[:joins] = normalize_joins(Array(value))
-      when :limit
-        normalized[:limit] = coerce_integer_min(value, :limit, 1)
-      when :offset
-        normalized[:offset] = coerce_integer_min(value, :offset, 0)
-      when :page
-        normalized[:page] = coerce_integer_min(value, :page, 1)
-      when :per_page
-        normalized[:per_page] = coerce_integer_min(value, :per, 1)
-      when :options
-        normalized[:options] = (value || {}).dup
-      when :grouping
-        normalized[:grouping] = normalize_grouping(value)
-      when :preset_name
-        normalized[:preset_name] = value&.to_s&.strip
-      when :preset_mode
-        normalized[:preset_mode] = value&.to_sym
-      when :curation
-        normalized[:curation] = normalize_curation_input(value)
-      when :facet_fields
-        normalized[:facet_fields] = Array(value).flatten.compact
-      when :facet_max_values
-        normalized[:facet_max_values] = Array(value).flatten.compact
-      when :facet_queries
-        normalized[:facet_queries] = Array(value).flatten.compact
-      when :highlight
-        normalized[:highlight] = normalize_highlight_input(value)
-      when :ranking
-        normalized[:ranking] = normalize_ranking_input(value || {})
-      when :hit_limits
-        normalized[:hit_limits] = normalize_hit_limits_input(value || {})
-      end
-    end
-
-    # One-time migration: when AST is empty and legacy string filters exist, map
-    # each fragment to AST::Raw and prefer AST going forward. Idempotent.
-    def migrate_legacy_filters_to_ast!(state)
-      return unless state.is_a?(Hash)
-
-      ast_nodes = Array(state[:ast]).flatten.compact
-      legacy = Array(state[:filters]).flatten.compact
-      return if !ast_nodes.empty? || legacy.empty?
-
-      raw_nodes = legacy.map { |fragment| SearchEngine::AST.raw(String(fragment)) }
-      state[:ast] = raw_nodes
-    end
-
-    # Normalize where arguments into an array of string fragments safe for Typesense.
-    # Supports hash, raw string, and template-with-placeholders.
-    def normalize_where(args)
-      list = Array(args).flatten.compact
-      return [] if list.empty?
-
-      fragments = []
-      i = 0
-      known_attrs = safe_attributes_map
-
-      while i < list.length
-        entry = list[i]
-        case entry
-        when Hash
-          validate_hash_keys!(entry, known_attrs)
-          fragments.concat(SearchEngine::Filters::Sanitizer.build_from_hash(entry, known_attrs))
-          i += 1
-        when String
-          i = normalize_where_process_string!(fragments, entry, list, i)
-        when Symbol
-          # Treat symbol as raw string fragment for compatibility
-          fragments << entry.to_s
-          i += 1
-        when Array
-          # Recurse over nested arrays
-          nested = normalize_where(entry)
-          fragments.concat(nested)
-          i += 1
-        else
-          raise ArgumentError, "unsupported where argument of type #{entry.class}"
-        end
-      end
-
-      fragments
-    end
-
-    def normalize_where_process_string!(fragments, entry, list, i)
-      if entry.match?(/(?<!\\)\?/) # has unescaped placeholders
-        tail = list[(i + 1)..] || []
-        needed = SearchEngine::Filters::Sanitizer.count_placeholders(entry)
-        args_for_template = tail.first(needed)
-        if args_for_template.length != needed
-          raise ArgumentError, "expected #{needed} args for #{needed} placeholders, got #{args_for_template.length}"
-        end
-
-        fragments << SearchEngine::Filters::Sanitizer.apply_placeholders(entry, args_for_template)
-        i + 1 + needed
-      else
-        fragments << entry.to_s
-        i + 1
-      end
-    end
-
-    # Parse and normalize order input into an array of "field:dir" strings.
-    def normalize_order(value) # rubocop:disable Metrics/AbcSize, Metrics/PerceivedComplexity
-      return [] if value.nil?
-
-      case value
-      when Hash
-        value.flat_map do |k, dir|
-          if dir.is_a?(Hash)
-            # Nested assoc: { assoc => { field => :asc|:desc } }
-            assoc = k.to_sym
-            # Validate assoc exists and join applied
-            @klass.join_for(assoc)
-            SearchEngine::Joins::Guard.ensure_join_applied!(joins_list, assoc, context: 'sorting')
-
-            dir.flat_map do |field_name, d|
-              field = field_name.to_s.strip
-              raise ArgumentError, 'order: field name must be non-empty' if field.empty?
-
-              begin
-                cfg = @klass.join_for(assoc)
-                SearchEngine::Joins::Guard.validate_joined_field!(cfg, field)
-              rescue StandardError
-                # best-effort only
-              end
-
-              direction = d.to_s.strip.downcase
-              unless %w[asc desc].include?(direction)
-                raise ArgumentError,
-                      "order: direction must be :asc or :desc (got #{d.inspect} for field #{field_name.inspect})"
-              end
-
-              "$#{assoc}.#{field}:#{direction}"
-            end
-          else
-            field = k.to_s.strip
-            raise ArgumentError, 'order: field name must be non-empty' if field.empty?
-
-            direction = dir.to_s.strip.downcase
-            unless %w[asc desc].include?(direction)
-              raise ArgumentError, "order: direction must be :asc or :desc (got #{dir.inspect} for field #{k.inspect})"
-            end
-
-            "#{field}:#{direction}"
-          end
-        end
-      when String
-        value.split(',').map(&:strip).reject(&:empty?).map do |chunk|
-          name, direction = chunk.split(':', 2).map { |p| p.to_s.strip }
-          if name.empty? || direction.empty?
-            raise ArgumentError, "order: expected 'field:direction' (got #{chunk.inspect})"
-          end
-
-          downcased = direction.downcase
-          unless %w[asc desc].include?(downcased)
-            raise ArgumentError,
-                  "order: direction must be :asc or :desc (got #{direction.inspect} for field #{name.inspect})"
-          end
-
-          "#{name}:#{downcased}"
-        end
-      when Array
-        # Allow arrays of accepted forms
-        value.flat_map { |v| normalize_order(v) }
-      when Symbol
-        # Back-compat: treat as ascending
-        field = value.to_s.strip
-        raise ArgumentError, 'order: field name must be non-empty' if field.empty?
-
-        ["#{field}:asc"]
-      else
-        raise ArgumentError, "order: unsupported input #{value.class}"
-      end
-    end
-
-    # Dedupe by field with last-wins semantics while preserving last positions.
-    def dedupe_orders_last_wins(list)
-      return [] if list.nil? || list.empty?
-
-      last_by_field = {}
-      list.each_with_index do |entry, idx|
-        field, dir = entry.split(':', 2)
-        last_by_field[field] = { idx: idx, str: "#{field}:#{dir}" }
-      end
-      last_by_field.values.sort_by { |h| h[:idx] }.map { |h| h[:str] }
-    end
-
-    # Base-only normalization used internally and for legacy callers.
-    def normalize_select(fields)
-      list = Array(fields).flatten.compact
-      return [] if list.empty?
-
-      known_attrs = safe_attributes_map
-      known = known_attrs.keys.map(&:to_s)
-
-      ordered = []
-      list.each do |f|
-        name = f.to_s.strip
-        raise ArgumentError, 'select: field names must be non-empty' if name.empty?
-
-        if !known.empty? && !known.include?(name)
-          suggestions = suggest_fields(name.to_sym, known_attrs.keys.map(&:to_sym))
-          suggest = if suggestions.empty?
-                      ''
-                    elsif suggestions.length == 1
-                      " (did you mean :#{suggestions.first}?)"
-                    else
-                      last = suggestions.last
-                      others = suggestions[0..-2].map { |s| ":#{s}" }.join(', ')
-                      " (did you mean #{others}, or :#{last}?)"
-                    end
-          raise SearchEngine::Errors::UnknownField,
-                "UnknownField: unknown field #{name.inspect} for #{klass_name_for_inspect}#{suggest}"
-        end
-
-        ordered << name unless ordered.include?(name)
-      end
-      ordered
-    end
-
-    # Extended normalization supporting nested association selections.
-    # Returns a Hash with keys: :base (Array<String>), :nested (Hash<Symbol,Array<String>>), :nested_order (Array<Symbol>).
-    def normalize_select_input(fields, context: 'selecting fields') # rubocop:disable Metrics/AbcSize
-      list = Array(fields).flatten.compact
-      return { base: [], nested: {}, nested_order: [] } if list.empty?
-
-      base = []
-      nested = {}
-      nested_order = []
-
-      add_base = ->(val) { base.concat(normalize_select([val])) }
-
-      add_nested = lambda do |assoc, values|
-        key = assoc.to_sym
-        @klass.join_for(key)
-        # Enforce that the relation has applied the join before selecting nested fields
-        SearchEngine::Joins::Guard.ensure_join_applied!(joins_list, key, context: context)
-
-        items = case values
-                when Array then values
-                when nil then []
-                else [values]
-                end
-        names = items.flatten.compact.map(&:to_s).map(&:strip).reject(&:empty?)
-        return if names.empty?
-
-        # Validate joined field names against target collection metadata
-        cfg = @klass.join_for(key)
-        names.each do |fname|
-          SearchEngine::Joins::Guard.validate_joined_field!(cfg, fname, source_klass: @klass)
-        end
-
-        existing = Array(nested[key])
-        merged = (existing + names).each_with_object([]) { |n, acc| acc << n unless acc.include?(n) }
-        nested[key] = merged
-        nested_order << key unless nested_order.include?(key)
-      end
-
-      i = 0
-      while i < list.length
-        entry = list[i]
-        case entry
-        when Hash
-          entry.each { |k, v| add_nested.call(k, v) }
-          i += 1
-        when Symbol, String
-          add_base.call(entry)
-          i += 1
-        when Array
-          inner = Array(entry).flatten.compact
-          inner.each { |el| list << el }
-          i += 1
-        else
-          raise SearchEngine::Errors::ConflictingSelection,
-                "ConflictingSelection: unsupported input #{entry.class} in selection"
-        end
-      end
-
-      { base: base.uniq, nested: nested, nested_order: nested_order }
-    end
-
-    def normalize_grouping(value)
-      return nil if value.nil? || value.empty?
-      raise ArgumentError, 'grouping: expected a Hash' unless value.is_a?(Hash)
-
-      field = value[:field]
-      limit = value[:limit]
-      missing_values = value[:missing_values]
-
-      unless field.is_a?(Symbol) || field.is_a?(String)
-        raise SearchEngine::Errors::InvalidGroup,
-              'InvalidGroup: field must be a Symbol or String'
-      end
-
-      # Disallow joined/path fields like "$assoc.field"
-      field_str = field.to_s
-      if field_str.start_with?('$') || field_str.include?('.')
-        raise SearchEngine::Errors::UnsupportedGroupField.new(
-          %(UnsupportedGroupField: grouping supports base fields only (got #{field_str.inspect})),
-          doc: 'docs/grouping.md#troubleshooting',
-          details: { field: field_str }
-        )
-      end
-
-      # Validate existence against declared attributes when available
-      attrs = safe_attributes_map
-      unless attrs.nil? || attrs.empty?
-        sym = field.to_sym
-        unless attrs.key?(sym)
-          msg = build_invalid_group_unknown_field_message(sym)
-          raise SearchEngine::Errors::InvalidGroup.new(
-            msg,
-            doc: 'docs/grouping.md#troubleshooting',
-            details: { field: sym }
-          )
-        end
-      end
-
-      # Validate limit positivity when provided
-      if !limit.nil? && !(limit.is_a?(Integer) && limit >= 1)
-        got = limit.nil? ? 'nil' : limit.inspect
-        raise SearchEngine::Errors::InvalidGroup.new(
-          "InvalidGroup: limit must be a positive integer (got #{got})",
-          doc: 'docs/grouping.md#troubleshooting',
-          details: { limit: limit }
-        )
-      end
-
-      # Validate missing_values strict boolean
-      unless [true, false].include?(missing_values)
-        raise SearchEngine::Errors::InvalidGroup.new(
-          "InvalidGroup: missing_values must be boolean (got #{missing_values.inspect})",
-          doc: 'docs/grouping.md#troubleshooting',
-          details: { missing_values: missing_values }
-        )
-      end
-
-      { field: field.to_sym, limit: limit, missing_values: missing_values }
-    end
-
-    def coerce_integer_min(value, name, min)
-      return nil if value.nil?
-
-      integer =
-        case value
-        when Integer then value
-        else Integer(value)
-        end
-
-      raise ArgumentError, "#{name} must be >= #{min}" if integer < min
-
-      integer
-    rescue ArgumentError, TypeError
-      raise ArgumentError, "#{name} must be an Integer or nil"
-    end
-
-    def coerce_boolean_strict(value, name)
-      case value
-      when true, false
-        value
-      when String
-        s = value.to_s.strip.downcase
-        return true  if %w[1 true yes on t].include?(s)
-
-        return false if %w[0 false no off f].include?(s)
-
-        raise ArgumentError, "#{name} must be a boolean"
-      when Integer
-        return true  if value == 1
-
-        return false if value.zero?
-
-        raise ArgumentError, "#{name} must be a boolean"
-      else
-        raise ArgumentError, "#{name} must be a boolean"
-      end
-    end
-
-    def deep_dup(obj)
-      case obj
-      when Hash
-        obj.transform_values(&method(:deep_dup))
-      when Array
-        obj.map(&method(:deep_dup))
-      else
-        obj
-      end
-    end
-
-    def deep_freeze_inplace(obj)
-      case obj
-      when Hash
-        obj.each_value { |v| deep_freeze_inplace(v) }
-        obj.freeze
-      when Array
-        obj.each { |el| deep_freeze_inplace(el) }
-        obj.freeze
-      else
-        obj.freeze if obj.is_a?(String)
-      end
-      obj
     end
 
     def safe_attributes_map
@@ -1833,7 +242,6 @@ module SearchEngine
         cfg = SearchEngine.config
         return unless cfg.respond_to?(:strict_fields) ? cfg.strict_fields : true
       rescue StandardError
-        # Intentionally ignore config access errors and fall back to strict behavior
       end
 
       klass_name = klass_name_for_inspect
@@ -1842,153 +250,23 @@ module SearchEngine
       raise ArgumentError, "Unknown attribute #{unknown_name} for #{klass_name}. Known: #{known_list}"
     end
 
-    # Build include_fields string with nested association segments first, then base fields.
-    def compile_include_fields_string
-      # Start from explicit selections
-      include_nested_map = @state[:select_nested] || {}
-      include_base = Array(@state[:select])
+    def collection_name_for_klass
+      return @klass.collection if @klass.respond_to?(:collection) && @klass.collection
 
-      # Apply excludes if present (only to trim explicit includes)
-      exclude_base = Array(@state[:exclude])
-      exclude_nested_map = @state[:exclude_nested] || {}
-
-      # Base: only emit when include list is non-empty
-      base_segment = include_base.empty? ? [] : (include_base - exclude_base)
-
-      # Nested: only emit for associations with explicit includes; trim by nested excludes
-      applied_joins = joins_list
-      nested_segments = []
-      Array(@state[:select_nested_order]).each do |assoc|
-        next unless applied_joins.include?(assoc)
-
-        inc_fields = Array(include_nested_map[assoc])
-        next if inc_fields.empty?
-
-        exc_fields = Array(exclude_nested_map[assoc])
-        fields = (inc_fields - exc_fields).map(&:to_s).reject(&:empty?)
-        # Ensure deterministic order within the assoc group
-        fields = fields.sort
-        next if fields.empty?
-
-        nested_segments << "$#{assoc}(#{fields.join(',')})"
+      begin
+        mapping = SearchEngine::Registry.mapping
+        found = mapping.find { |(_, kls)| kls == @klass }
+        return found.first if found
+      rescue StandardError
       end
 
-      segments = []
-      segments.concat(nested_segments)
-      segments.concat(base_segment) unless base_segment.empty?
-
-      segments.join(',')
+      raise ArgumentError, "Unknown collection for #{klass_name_for_inspect}"
     end
 
-    # Build exclude_fields string with nested association segments first, then base fields.
-    # Emitted only for parts we cannot derive into include_fields (e.g., no attributes available).
-    def compile_exclude_fields_string
-      exclude_nested_order = Array(@state[:exclude_nested_order])
-      exclude_nested_map = @state[:exclude_nested] || {}
-      exclude_base = Array(@state[:exclude])
-
-      # If includes are present for base, excludes are already applied via include_fields
-      include_base = Array(@state[:select])
-      base_part = include_base.empty? ? exclude_base : []
-
-      segments = []
-
-      # Nested excludes: only emit when there is no explicit include for that assoc
-      include_nested_map = @state[:select_nested] || {}
-
-      # Preserve order by first mention in excludes (stable)
-      exclude_nested_order.each do |assoc|
-        next if Array(include_nested_map[assoc]).any?
-
-        fields = Array(exclude_nested_map[assoc]).map(&:to_s).reject(&:empty?)
-        # Ensure deterministic order within the assoc group
-        fields = fields.sort
-        next if fields.empty?
-
-        segments << "$#{assoc}(#{fields.join(',')})"
-      end
-
-      segments.concat(base_part) unless base_part.empty?
-      segments.join(',')
+    def client
+      @__client ||= (SearchEngine.config.respond_to?(:client) && SearchEngine.config.client) || SearchEngine::Client.new # rubocop:disable Naming/MemoizedInstanceVariableName
     end
 
-    # Ensure the relation has executed the search and memoized the Result.
-    # Thread-safe: double-checked locking around the first load.
-    # @return [void]
-    def ensure_loaded!
-      return if @__loaded && @__result_memo
-
-      @__load_lock.synchronize do
-        return if @__loaded && @__result_memo
-
-        execute
-      end
-
-      nil
-    end
-
-    # Execute the search via the client and memoize the Result.
-    #
-    # Compiles body params and derives URL/common options by starting from
-    # configuration defaults and applying relation-level overrides (whitelisted).
-    # Instrumentation is performed by the client to avoid duplicate events.
-    #
-    # @return [SearchEngine::Result]
-    def execute
-      collection = collection_name_for_klass
-      params = to_typesense_params
-
-      # Start from config defaults, then apply relation-level overrides
-      url_opts = ClientOptions.url_options_from_config(SearchEngine.config)
-      overrides = build_url_opts
-      url_opts.merge!(overrides) unless overrides.empty?
-
-      result = client.search(collection: collection, params: params, url_opts: url_opts)
-      # Wrap with selection/facets context for hydration strictness and facet labels
-      selection_ctx = build_selection_context
-      facets_ctx = build_facets_context
-      if selection_ctx || facets_ctx
-        result = SearchEngine::Result.new(result.raw, klass: @klass, selection: selection_ctx, facets: facets_ctx)
-      end
-      # One-shot validator, executed once per materialization
-      enforce_hit_validator_if_needed!(result.found, collection: collection)
-      @__result_memo = result
-      @__loaded = true
-      result
-    end
-
-    # Perform a minimal request to obtain only the total `found` count.
-    # Does not memoize the full Result.
-    # @return [Integer]
-    def fetch_found_only
-      collection = collection_name_for_klass
-      base = to_typesense_params
-
-      minimal = base.dup
-      minimal[:per_page] = 1
-      minimal[:page] = 1
-      # Keep include_fields minimal to reduce payload
-      minimal[:include_fields] = 'id'
-
-      # Merge config defaults with any relation-level overrides
-      url_opts = ClientOptions.url_options_from_config(SearchEngine.config)
-      overrides = build_url_opts
-      url_opts.merge!(overrides) unless overrides.empty?
-
-      result = client.search(collection: collection, params: minimal, url_opts: url_opts)
-      count = result.found.to_i
-      enforce_hit_validator_if_needed!(count, collection: collection)
-      count
-    end
-
-    # Compile relation state into Typesense search params.
-    # @return [Hash]
-    def build_search_params
-      to_typesense_params
-    end
-
-    # Derive URL/common options for the client request.
-    # @return [Hash]
     def build_url_opts
       opts = @state[:options] || {}
       url = {}
@@ -2003,327 +281,8 @@ module SearchEngine
       url.compact
     end
 
-    # Compute page and per_page from explicit page/per or limit/offset fallback.
-    # Omits pagination keys entirely when nothing is specified.
-    # @return [Hash{Symbol=>Integer}]
-    def compute_pagination
-      page = @state[:page]
-      per = @state[:per_page]
-
-      if page || per
-        out = {}
-        if per && page
-          out[:page] = page
-          out[:per_page] = per
-        elsif per && !page
-          out[:page] = 1
-          out[:per_page] = per
-        elsif page && !per
-          out[:page] = page
-        end
-        return out
-      elsif @state[:limit]
-        limit = @state[:limit]
-        off = @state[:offset] || 0
-        computed_page = (off.to_i / limit.to_i) + 1
-        return { page: computed_page, per_page: limit }
-      end
-
-      {}
-    end
-
-    # Resolve the Typesense collection name for the bound class.
-    # @return [String]
-    def collection_name_for_klass
-      return @klass.collection if @klass.respond_to?(:collection) && @klass.collection
-
-      # Fallback: reverse-lookup in registry
-      begin
-        mapping = SearchEngine::Registry.mapping
-        found = mapping.find { |(_, kls)| kls == @klass }
-        return found.first if found
-      rescue StandardError
-        # ignore
-      end
-
-      raise ArgumentError, "Unknown collection for #{klass_name_for_inspect}"
-    end
-
-    def client
-      @__client ||= (SearchEngine.config.respond_to?(:client) && SearchEngine.config.client) || SearchEngine::Client.new # rubocop:disable Naming/MemoizedInstanceVariableName
-    end
-
-    def option_value(hash, key)
-      if hash.key?(key)
-        hash[key]
-      else
-        hash[key.to_s]
-      end
-    end
-
-    def truncate_for_inspect(str, max = 80)
-      return str unless str.is_a?(String)
-      return str if str.length <= max
-
-      "#{str[0, max]}..."
-    end
-
-    def add_pagination_line!(lines, params)
-      page = params[:page]
-      per = params[:per_page]
-      return unless page || per
-
-      if page && per
-        lines << "  page/per: #{page}/#{per}"
-      elsif page
-        lines << "  page/per: #{page}/"
-      elsif per
-        lines << "  page/per: /#{per}"
-      end
-    end
-
-    def append_selection_inspect_parts(parts, compiled)
-      selected_len = Array(@state[:select]).length
-      parts << "select=#{selected_len}" unless selected_len.zero?
-
-      inc_str = compiled[:include_fields]
-      parts << %(sel="#{truncate_for_inspect(inc_str)}") if inc_str && !inc_str.to_s.empty?
-      exc_str = compiled[:exclude_fields]
-      parts << %(xsel="#{truncate_for_inspect(exc_str)}") if exc_str && !exc_str.to_s.empty?
-    end
-
-    def append_selection_explain_lines(lines, params)
-      if params[:include_fields] && !params[:include_fields].to_s.strip.empty?
-        lines << "  select: #{params[:include_fields]}"
-      end
-      return lines unless params[:exclude_fields] && !params[:exclude_fields].to_s.strip.empty?
-
-      lines << "  exclude: #{params[:exclude_fields]}"
-    end
-
-    def append_curation_explain_lines(lines)
-      cur = @state[:curation]
-      return lines unless cur
-
-      pinned = Array(cur[:pinned]).map(&:to_s).reject(&:empty?)
-      hidden = Array(cur[:hidden]).map(&:to_s).reject(&:empty?)
-      tags   = Array(cur[:override_tags]).map(&:to_s).reject(&:empty?)
-      fch    = cur[:filter_curated_hits]
-
-      lines << "  Pinned: #{pinned.join(', ')}" unless pinned.empty?
-      lines << "  Hidden: #{hidden.join(', ')}" unless hidden.empty?
-      lines << "  Override tags: #{tags.join(', ')}" unless tags.empty?
-      lines << "  Filter curated hits: #{fch}" unless fch.nil?
-      lines
-    end
-
-    def add_effective_selection_tokens!(lines)
-      include_root = Array(@state[:select]).map(&:to_s)
-      exclude_root = Array(@state[:exclude]).map(&:to_s)
-      return if include_root.empty? && exclude_root.empty?
-
-      effective = include_root.empty? ? include_root : (include_root - exclude_root)
-      parts = ['selection:']
-      parts << "sel=#{effective.join(',')}" if effective.any?
-      parts << "xsel=#{exclude_root.join(',')}" if exclude_root.any?
-      lines << "  #{parts.join(' ')}"
-    end
-
-    # Normalize and validate join names, preserving order and duplicates.
-    def normalize_joins(values)
-      list = Array(values).flatten.compact
-      return [] if list.empty?
-
-      names = list.map do |v|
-        case v
-        when Symbol, String
-          v.to_sym
-        else
-          raise ArgumentError, "joins: expected symbols/strings (got #{v.class})"
-        end
-      end
-
-      # Validate against declared joins; rely on join_for to raise with suggestions
-      names.each { |name| @klass.join_for(name) }
-      names
-    end
-
-    # Compile filter_by string from AST nodes or legacy fragments.
-    # @param ast_nodes [Array<SearchEngine::AST::Node>]
-    # @return [String, nil]
-    def compiled_filter_by(ast_nodes)
-      unless ast_nodes.empty?
-        compiled = SearchEngine::Compiler.compile(ast_nodes, klass: @klass)
-        return nil if compiled.to_s.empty?
-
-        return compiled
-      end
-
-      fragments = Array(@state[:filters])
-      return nil if fragments.empty?
-
-      fragments.join(' && ')
-    end
-
-    # Compile sort_by from normalized order entries.
-    # @param orders [Array<String>]
-    # @return [String, nil]
-    def compiled_sort_by(orders)
-      list = Array(orders)
-      return nil if list.empty?
-
-      list.join(',')
-    end
-
-    # Build a JSON-serializable join context for Typesense.
-    # Internal, documented shape used by downstream layers for diagnostics and planning:
-    #   {
-    #     assocs: [:authors, ...],
-    #     fields_by_assoc: { authors: ["first_name", ...] },
-    #     referenced_in: { include: [:authors], filter: [:authors], sort: [:authors] }
-    #   }
-    # @param ast_nodes [Array<SearchEngine::AST::Node>]
-    # @param orders [Array<String>]
-    # @return [Hash]
-    def build_join_context(ast_nodes:, orders:)
-      applied = Array(@state[:joins])
-      return {} if applied.empty?
-
-      # Dedupe while preserving first occurrence order
-      assocs = []
-      applied.each { |a| assocs << a unless assocs.include?(a) }
-
-      nested_map = @state[:select_nested] || {}
-      nested_order = Array(@state[:select_nested_order])
-
-      fields_by_assoc = {}
-      assocs.each do |assoc|
-        fields = Array(nested_map[assoc]).map(&:to_s).reject(&:empty?)
-        fields_by_assoc[assoc] = fields unless fields.empty?
-      end
-
-      include_refs = nested_order.select { |a| Array(nested_map[a]).any? }
-      filter_refs = extract_assocs_from_ast(ast_nodes)
-      sort_refs = extract_assocs_from_orders(orders)
-
-      referenced_in = {}
-      referenced_in[:include] = include_refs unless include_refs.empty?
-      referenced_in[:filter] = filter_refs unless filter_refs.empty?
-      referenced_in[:sort] = sort_refs unless sort_refs.empty?
-
-      out = {}
-      out[:assocs] = assocs unless assocs.empty?
-      out[:fields_by_assoc] = fields_by_assoc unless fields_by_assoc.empty?
-      out[:referenced_in] = referenced_in unless referenced_in.empty?
-      out
-    end
-
-    # Walk AST nodes and collect association names used via "$assoc.field" LHS.
-    # @param nodes [Array<SearchEngine::AST::Node>]
-    # @return [Array<Symbol>] unique assoc names in first-mention order
-    def extract_assocs_from_ast(nodes)
-      list = Array(nodes).flatten.compact
-      return [] if list.empty?
-
-      seen = []
-      walker = lambda do |node|
-        return unless node.is_a?(SearchEngine::AST::Node)
-
-        if node.respond_to?(:field)
-          field = node.field.to_s
-          if field.start_with?('$')
-            m = field.match(/^\$(\w+)\./)
-            if m
-              name = m[1].to_sym
-              seen << name unless seen.include?(name)
-            end
-          end
-        end
-
-        Array(node.children).each { |child| walker.call(child) }
-      end
-
-      list.each { |n| walker.call(n) }
-      seen
-    end
-
-    # Parse order strings and collect assoc names used via "$assoc.field:dir".
-    # @param orders [Array<String>]
-    # @return [Array<Symbol>] unique assoc names in first-mention order
-    def extract_assocs_from_orders(orders)
-      list = Array(orders).flatten.compact
-      return [] if list.empty?
-
-      seen = []
-      list.each do |entry|
-        field, _dir = entry.to_s.split(':', 2)
-        next unless field&.start_with?('$')
-
-        m = field.match(/^\$(\w+)\./)
-        next unless m
-
-        name = m[1].to_sym
-        seen << name unless seen.include?(name)
-      end
-      seen
-    end
-
-    # Build an actionable InvalidGroup message for unknown field with suggestions.
-    def build_invalid_group_unknown_field_message(field_sym)
-      klass_name = klass_name_for_inspect
-      known = safe_attributes_map.keys.map(&:to_sym)
-      suggestions = suggest_fields(field_sym, known)
-      suggestion_str =
-        if suggestions.empty?
-          ''
-        elsif suggestions.length == 1
-          " (did you mean :#{suggestions.first}?)"
-        else
-          last = suggestions.last
-          others = suggestions[0..-2].map { |s| ":#{s}" }.join(', ')
-          " (did you mean #{others}, or :#{last}?)"
-        end
-      "InvalidGroup: unknown field :#{field_sym} for grouping on #{klass_name}#{suggestion_str}"
-    end
-
-    # Lightweight suggestion helper using Levenshtein; returns up to 3 candidates within distance <= 2.
-    def suggest_fields(field_sym, known_syms)
-      return [] if known_syms.nil? || known_syms.empty?
-
-      input = field_sym.to_s
-      candidates = known_syms.map(&:to_s)
-      begin
-        require 'did_you_mean'
-        require 'did_you_mean/levenshtein'
-      rescue StandardError
-        return []
-      end
-
-      distances = candidates.each_with_object({}) do |cand, acc|
-        acc[cand] = DidYouMean::Levenshtein.distance(input, cand)
-      end
-      # Select up to 3 with minimal distance under threshold
-      sorted = distances.sort_by { |(_cand, d)| d }
-      threshold = 2
-      sorted.take(3).select { |(_cand, d)| d <= threshold }.map { |cand, _d| cand.to_sym }
-    end
-
-    def grouping_warnings_enabled?(cfg)
-      cfg&.grouping&.warn_on_ambiguous
-    end
-
-    def excludes_nulls_for_field?(ast_nodes, field_str)
-      field = field_str.to_s
-      Array(ast_nodes).flatten.any? do |node|
-        case node
-        when SearchEngine::AST::NotEq
-          node.field.to_s == field && node.value.nil?
-        when SearchEngine::AST::NotIn
-          node.field.to_s == field && Array(node.values).include?(nil)
-        else
-          false
-        end
-      end
+    def selection_maps_all_empty?(map)
+      Array(map.values).all? { |v| Array(v).empty? }
     end
 
     def build_selection_context # rubocop:disable Metrics/PerceivedComplexity
@@ -2336,7 +295,6 @@ module SearchEngine
       nothing_excluded = exclude_base.empty? && selection_maps_all_empty?(exclude_nested)
       return nil if nothing_selected && nothing_excluded
 
-      # Compute effective selection (exclude wins)
       effective_base = include_base.empty? ? nil : (include_base - exclude_base).map(&:to_s).reject(&:empty?)
 
       nested_effective = {}
@@ -2356,202 +314,6 @@ module SearchEngine
       selection unless selection.empty?
     end
 
-    def selection_maps_all_empty?(map)
-      Array(map.values).all? { |v| Array(v).empty? }
-    end
-
-    def append_preset_explain_line(lines, params)
-      return lines unless @state[:preset_name]
-
-      mode = @state[:preset_mode] || :merge
-      if (conf = Array(params[:_preset_conflicts])) && !conf.empty?
-        keys = conf.map(&:to_s).sort
-        lines << "  preset: #{@state[:preset_name]} (mode=#{mode} dropped: #{keys.join(',')})"
-      else
-        lines << "  preset: #{@state[:preset_name]} (mode=#{mode})"
-      end
-      lines
-    end
-
-    # Instrument preset conflicts in :lock mode without affecting compile flow.
-    # Swallows any observer errors to keep compilation pure/deterministic.
-    # @param mode [Symbol]
-    # @param name [String]
-    # @param conflicts [Array<Symbol>]
-    # @return [void]
-    def instrument_preset_conflicts(mode, name, conflicts)
-      return unless defined?(SearchEngine::Instrumentation)
-      return if Array(conflicts).empty?
-
-      payload = {
-        keys: Array(conflicts).map(&:to_sym).sort,
-        mode: mode,
-        preset_name: name,
-        count: Array(conflicts).size
-      }
-      SearchEngine::Instrumentation.instrument('search_engine.preset.conflict', payload) {}
-    rescue StandardError
-      nil
-    end
-
-    # Curation normalizers
-    # ---------------------
-    def normalize_curation_ids(values)
-      list = Array(values).flatten(1).compact
-      list.map { |v| v.to_s.strip }.reject(&:empty?)
-    end
-
-    def normalize_curation_tags(values)
-      list = Array(values).flatten(1).compact.map { |v| v.to_s.strip }.reject(&:empty?)
-      list.each_with_object([]) { |t, acc| acc << t unless acc.include?(t) }
-    end
-
-    def normalize_curation_input(value)
-      return nil if value.nil? || (value.respond_to?(:empty?) && value.empty?)
-      raise ArgumentError, 'curation must be a Hash' unless value.is_a?(Hash)
-
-      pinned = normalize_curation_ids(value[:pinned] || value['pinned'])
-      hidden = normalize_curation_ids(value[:hidden] || value['hidden'])
-      tags = normalize_curation_tags(value[:override_tags] || value['override_tags'])
-
-      raw_fch = (value.key?(:filter_curated_hits) ? value[:filter_curated_hits] : value['filter_curated_hits'])
-      fch = raw_fch.nil? ? nil : coerce_boolean_strict(raw_fch, :filter_curated_hits)
-
-      { pinned: pinned, hidden: hidden, override_tags: tags, filter_curated_hits: fch }
-    end
-
-    def curated_indices_for_current_result
-      @__result_memo.to_a.each_with_index.select do |obj, _idx|
-        obj.respond_to?(:curated_hit?) && obj.curated_hit?
-      end.map(&:last)
-    end
-
-    def curation_filter_curated_hits?
-      @state[:curation] && @state[:curation][:filter_curated_hits]
-    end
-
-    # Faceting DSL
-    # ------------
-    # @see docs/faceting.md
-    def facet_by(field, max_values: nil, sort: nil, stats: nil) # rubocop:disable Metrics/AbcSize
-      name = field.to_s.strip
-      raise SearchEngine::Errors::InvalidParams, 'facet_by: field name must be non-empty' if name.empty?
-
-      if name.start_with?('$') || name.include?('.')
-        raise SearchEngine::Errors::InvalidParams.new(
-          %(facet_by: supports base fields only (got #{name.inspect})),
-          doc: 'docs/faceting.md#supported-options',
-          details: { field: name }
-        )
-      end
-
-      attrs = safe_attributes_map
-      unless attrs.nil? || attrs.empty? || attrs.key?(name.to_sym)
-        suggestions = suggest_fields(name.to_sym, attrs.keys.map(&:to_sym))
-        suggest = if suggestions.empty?
-                    ''
-                  elsif suggestions.length == 1
-                    " (did you mean :#{suggestions.first}?)"
-                  else
-                    last = suggestions.last
-                    others = suggestions[0..-2].map { |s| ":#{s}" }.join(', ')
-                    " (did you mean #{others}, or :#{last}?)"
-                  end
-        raise SearchEngine::Errors::UnknownField,
-              "UnknownField: unknown field #{name.inspect} for #{klass_name_for_inspect}#{suggest}"
-      end
-
-      unless sort.nil?
-        raise SearchEngine::Errors::InvalidParams.new(
-          "facet_by: option :sort is not supported by Typesense facets (got #{sort.inspect})",
-          hint: 'Supported: default count-desc only at present.',
-          doc: 'docs/faceting.md#supported-options',
-          details: { sort: sort }
-        )
-      end
-
-      unless stats.nil?
-        raise SearchEngine::Errors::InvalidParams.new(
-          'facet_by: option :stats is not supported at present',
-          doc: 'docs/faceting.md#supported-options',
-          details: { stats: stats }
-        )
-      end
-
-      cap = nil
-      unless max_values.nil?
-        begin
-          cap = Integer(max_values)
-        rescue ArgumentError, TypeError
-          raise SearchEngine::Errors::InvalidParams, 'facet_by: max_values must be an Integer or nil'
-        end
-        raise SearchEngine::Errors::InvalidParams, 'facet_by: max_values must be >= 1' if cap < 1
-      end
-
-      spawn do |s|
-        fields = Array(s[:facet_fields])
-        s[:facet_fields] = fields.include?(name) ? fields : (fields + [name])
-
-        caps = Array(s[:facet_max_values])
-        s[:facet_max_values] = cap.nil? ? caps : (caps + [cap])
-      end
-    end
-
-    def facet_query(field, expression, label: nil) # rubocop:disable Metrics/AbcSize, Metrics/PerceivedComplexity
-      name = field.to_s.strip
-      raise SearchEngine::Errors::InvalidParams, 'facet_query: field name must be non-empty' if name.empty?
-
-      if name.start_with?('$') || name.include?('.')
-        raise SearchEngine::Errors::InvalidParams.new(
-          %(facet_query: supports base fields only (got #{name.inspect})),
-          doc: 'docs/faceting.md#supported-options',
-          details: { field: name }
-        )
-      end
-
-      attrs = safe_attributes_map
-      unless attrs.nil? || attrs.empty? || attrs.key?(name.to_sym)
-        suggestions = suggest_fields(name.to_sym, attrs.keys.map(&:to_sym))
-        suggest = if suggestions.empty?
-                    ''
-                  elsif suggestions.length == 1
-                    " (did you mean :#{suggestions.first}?)"
-                  else
-                    last = suggestions.last
-                    others = suggestions[0..-2].map { |s| ":#{s}" }.join(', ')
-                    " (did you mean #{others}, or :#{last}?)"
-                  end
-        raise SearchEngine::Errors::UnknownField,
-              "UnknownField: unknown field #{name.inspect} for #{klass_name_for_inspect}#{suggest}"
-      end
-
-      expr = expression.to_s.strip
-      raise SearchEngine::Errors::InvalidParams, 'facet_query: expression must be a non-empty String' if expr.empty?
-
-      if expr.include?('[') ^ expr.include?(']')
-        raise SearchEngine::Errors::InvalidParams.new(
-          %(facet_query: invalid range syntax #{expr.inspect} (unbalanced brackets)),
-          hint: 'Use shapes like "[0..9]", "[10..19]"',
-          doc: 'docs/faceting.md#facet-query-expressions',
-          details: { expr: expr }
-        )
-      end
-
-      label_str = label&.to_s&.strip
-
-      spawn do |s|
-        queries = Array(s[:facet_queries])
-        rec = { field: name, expr: expr }
-        rec[:label] = label_str unless label_str.nil? || label_str.empty?
-        exists = queries.any? { |q| q[:field] == rec[:field] && q[:expr] == rec[:expr] && q[:label] == rec[:label] }
-        s[:facet_queries] = exists ? queries : (queries + [rec])
-      end
-    end
-
-    public :facet_by, :facet_query
-
-    protected
-
     def build_facets_context
       fields = Array(@state[:facet_fields]).map(&:to_s)
       queries = Array(@state[:facet_queries]).map do |q|
@@ -2564,216 +326,18 @@ module SearchEngine
       { fields: fields.freeze, queries: queries.freeze }.freeze
     end
 
-    def build_highlight_context
-      h = @state[:highlight]
-      return nil unless h
+    # pluck helpers reside in Materializers
 
-      out = {}
-      f = Array(h[:fields]).map(&:to_s).reject(&:empty?)
-      out[:fields] = f unless f.empty?
-      ff = Array(h[:full_fields]).map(&:to_s).reject(&:empty?)
-      out[:full_fields] = ff unless ff.empty?
-      out[:start_tag] = h[:start_tag] if h[:start_tag]
-      out[:end_tag] = h[:end_tag] if h[:end_tag]
-      out[:affix_tokens] = h[:affix_tokens] unless h[:affix_tokens].nil?
-      out[:snippet_threshold] = h[:snippet_threshold] unless h[:snippet_threshold].nil?
-      out
+    def curated_indices_for_current_result
+      @__result_memo.to_a.each_with_index.select do |obj, _idx|
+        obj.respond_to?(:curated_hit?) && obj.curated_hit?
+      end.map(&:last)
     end
 
-    # Control usage of synonyms at query time.
-    #
-    # @param value [true,false,nil] when nil, clears the override (inherit server default)
-    # @return [SearchEngine::Relation]
-    # @since 0.1.0
-    # @example
-    #   SearchEngine::Product.where(q: "red color").use_synonyms(true)
-    def use_synonyms(value)
-      v = value.nil? ? nil : coerce_boolean_strict(value, :use_synonyms)
-      spawn do |s|
-        s[:use_synonyms] = v
-      end
+    def curation_filter_curated_hits?
+      @state[:curation] && @state[:curation][:filter_curated_hits]
     end
 
-    # Control usage of stopwords at query time.
-    #
-    # @param value [true,false,nil] when nil, clears the override (inherit server default)
-    # @return [SearchEngine::Relation]
-    # @since 0.1.0
-    # @example
-    #   SearchEngine::Product.where(q: "red color").use_stopwords(false)
-    def use_stopwords(value)
-      v = value.nil? ? nil : coerce_boolean_strict(value, :use_stopwords)
-      spawn do |s|
-        s[:use_stopwords] = v
-      end
-    end
-
-    def normalize_highlight_input(value)
-      h = value || {}
-      raise SearchEngine::Errors::InvalidOption, 'highlight must be a Hash of options' unless h.is_a?(Hash)
-
-      fields = Array(h[:fields] || h['fields']).flatten.compact.map { |f| f.to_s.strip }.reject(&:empty?)
-      full_fields = Array(h[:full_fields] || h['full_fields']).flatten.compact.map { |f| f.to_s.strip }.reject(&:empty?)
-      start_tag = h[:start_tag] || h['start_tag']
-      end_tag = h[:end_tag] || h['end_tag']
-      affix = h.key?(:affix_tokens) ? h[:affix_tokens] : h['affix_tokens']
-      snippet = h.key?(:snippet_threshold) ? h[:snippet_threshold] : h['snippet_threshold']
-
-      affix = nil if affix.nil?
-      affix = coerce_integer_min(affix, :highlight_affix_num_tokens, 0) unless affix.nil?
-      snippet = nil if snippet.nil?
-      snippet = coerce_integer_min(snippet, :highlight_snippet_threshold, 0) unless snippet.nil?
-
-      {
-        fields: fields,
-        full_fields: full_fields,
-        start_tag: start_tag&.to_s,
-        end_tag: end_tag&.to_s,
-        affix_tokens: affix,
-        snippet_threshold: snippet
-      }
-    end
-
-    # Normalize ranking DSL inputs into a compact, frozen Hash.
-    # Validates domains and coerces types; does not validate field membership for weights.
-    # That validation is deferred to compile when the effective query_by is known.
-    def normalize_ranking_input(value) # rubocop:disable Metrics/AbcSize, Metrics/MethodLength, Metrics/PerceivedComplexity
-      h = value || {}
-      unless h.is_a?(Hash)
-        raise SearchEngine::Errors::InvalidOption.new(
-          'InvalidOption: ranking expects a Hash of options',
-          hint: 'Use ranking(num_typos: 1, drop_tokens_threshold: 0.2,'\
-                'prioritize_exact_match: true, query_by_weights: { name: 2 })',
-          doc: 'docs/ranking.md#options'
-        )
-      end
-
-      out = {}
-
-      if h.key?(:num_typos) || h.key?('num_typos')
-        raw = h[:num_typos] || h['num_typos']
-        unless raw.nil?
-          begin
-            iv = Integer(raw)
-            unless [0, 1, 2].include?(iv)
-              raise SearchEngine::Errors::InvalidOption.new(
-                "InvalidOption: num_typos must be 0, 1, or 2 (got #{raw.inspect})",
-                doc: 'docs/ranking.md#options'
-              )
-            end
-            out[:num_typos] = iv
-          rescue ArgumentError, TypeError
-            raise SearchEngine::Errors::InvalidOption.new(
-              "InvalidOption: num_typos must be an Integer in {0,1,2} (got #{raw.inspect})",
-              doc: 'docs/ranking.md#options'
-            )
-          end
-        end
-      end
-
-      if h.key?(:drop_tokens_threshold) || h.key?('drop_tokens_threshold')
-        raw = h[:drop_tokens_threshold] || h['drop_tokens_threshold']
-        unless raw.nil?
-          begin
-            fv = Float(raw)
-            unless fv >= 0.0 && fv <= 1.0 && fv.finite?
-              raise SearchEngine::Errors::InvalidOption.new(
-                "InvalidOption: drop_tokens_threshold must be a float between 0.0 and 1.0 (got #{raw.inspect})",
-                doc: 'docs/ranking.md#options'
-              )
-            end
-            out[:drop_tokens_threshold] = fv
-          rescue ArgumentError, TypeError
-            raise SearchEngine::Errors::InvalidOption.new(
-              "InvalidOption: drop_tokens_threshold must be a float between 0.0 and 1.0 (got #{raw.inspect})",
-              doc: 'docs/ranking.md#options'
-            )
-          end
-        end
-      end
-
-      if h.key?(:prioritize_exact_match) || h.key?('prioritize_exact_match')
-        raw = h[:prioritize_exact_match] || h['prioritize_exact_match']
-        out[:prioritize_exact_match] = raw.nil? ? nil : coerce_boolean_strict(raw, :prioritize_exact_match)
-      end
-
-      if h.key?(:query_by_weights) || h.key?('query_by_weights')
-        raw = h[:query_by_weights] || h['query_by_weights']
-        unless raw.nil?
-          unless raw.is_a?(Hash)
-            raise SearchEngine::Errors::InvalidOption.new(
-              'InvalidOption: query_by_weights must be a Hash of { field => Integer }',
-              doc: 'docs/ranking.md#weights'
-            )
-          end
-          normalized = {}
-          raw.each do |k, v|
-            key = k.to_s.strip
-            next if key.empty?
-
-            begin
-              w = Integer(v)
-            rescue ArgumentError, TypeError
-              raise SearchEngine::Errors::InvalidOption.new(
-                "InvalidOption: weight for #{k.inspect} must be an Integer >= 0",
-                doc: 'docs/ranking.md#weights',
-                details: { field: k, weight: v }
-              )
-            end
-            if w.negative?
-              raise SearchEngine::Errors::InvalidOption.new(
-                "InvalidOption: weight for #{k.inspect} must be >= 0",
-                doc: 'docs/ranking.md#weights',
-                details: { field: k, weight: v }
-              )
-            end
-            normalized[key] = w
-          end
-          out[:query_by_weights] = normalized
-        end
-      end
-
-      out
-    end
-
-    # Normalize hit limits input into a compact Hash.
-    # Accepts keys :early_limit and :max and coerces to positive Integers when possible.
-    # @param value [Hash]
-    # @return [Hash]
-    def normalize_hit_limits_input(value)
-      unless value.is_a?(Hash)
-        raise SearchEngine::Errors::InvalidOption.new(
-          'InvalidOption: hit_limits expects a Hash of options',
-          doc: 'docs/hit_limits.md'
-        )
-      end
-
-      out = {}
-      if value.key?(:early_limit) || value.key?('early_limit')
-        raw = value[:early_limit] || value['early_limit']
-        begin
-          iv = Integer(raw)
-          out[:early_limit] = iv if iv.positive?
-        rescue StandardError
-          # ignore; validation occurs in the public chainers
-        end
-      end
-      if value.key?(:max) || value.key?('max')
-        raw = value[:max] || value['max']
-        begin
-          iv = Integer(raw)
-          out[:max] = iv if iv.positive?
-        rescue StandardError
-          # ignore; validation occurs in the public chainers
-        end
-      end
-      out
-    end
-
-    # Raise when total hits exceed the configured validator ceiling.
-    # @param total_hits [Integer]
-    # @param collection [String, nil]
-    # @return [void]
     def enforce_hit_validator_if_needed!(total_hits, collection: nil)
       hl = @state[:hit_limits]
       return unless hl && hl[:max]
